+++ R/as_dribble.R
@@ -49,7 +49,7 @@ as_dribble.dribble <- function(x, ...) x
 as_dribble.default <- function(x, ...) {
   stop_glue_data(
     list(x = glue_collapse(class(x), sep = "/")),
-    "Don't know how to coerce object of class {x} into a dribble"
+    "Don't know how to coerce object of class <{x}> into a dribble"
   )
 }