+++ R/drive_ls.R
@@ -1,13 +1,13 @@
-#' List contents of a folder or Team Drive
+#' List contents of a folder or shared drive
 #'
-#' List the contents of a folder or Team Drive, recursively or not. This is a
+#' List the contents of a folder or shared drive, recursively or not. This is a
 #' thin wrapper around [drive_find()], that simply adds one constraint: the
 #' search is limited to direct or indirect children of `path`.
 #'
 #' @param path Specifies a single folder on Google Drive whose contents you want
 #'   to list. Can be an actual path (character), a file id or URL marked with
-#'   [as_id()], or a [`dribble`]. If it is a Team Drive or is a folder on a Team
-#'   Drive, it must be passed as a [`dribble`].
+#'   [as_id()], or a [`dribble`]. If it is a shared drive or is a folder on a
+#'   shared drive, it must be passed as a [`dribble`].
 #' @param ... Any parameters that are valid for [drive_find()].
 #' @param recursive Logical, indicating if you want only direct children of
 #'   `path` (`recursive = FALSE`, the default) or all children, including
@@ -43,14 +43,8 @@ drive_ls <- function(path = NULL, ..., recursive = FALSE) {
   path <- confirm_single_file(path)
 
   params <- rlang::list2(...)
-  if (is_team_drivy(path)) {
-    if (is_team_drive(path)) {
-      params[["team_drive"]] <- as_id(path)
-    } else {
-      params[["team_drive"]] <- as_id(
-        path[["drive_resource"]][[1]][["teamDriveId"]]
-      )
-    }
+  if (is_shared_drive(path)) {
+    params[["shared_drive"]] <- as_id(path)
   }
 
   parent <- path[["id"]]