+++ R/team_drive.R
@@ -1,0 +1,79 @@
+#' Deprecated Team Drive functions
+#'
+#' @description
+#' `r lifecycle::badge('deprecated')`
+#' @template team-drives-description
+#'
+#' @inheritParams shared_drive_find
+#' @inheritParams shared_drive_get
+#' @inheritParams as_shared_drive
+#' @inheritParams is_shared_drive
+#' @template team_drive-plural
+#'
+#' @template dribble-return
+#'
+#' @keywords internal
+#' @name deprecated-team-drive-functions
+NULL
+
+#' @export
+#' @rdname deprecated-team-drive-functions
+team_drive_find <- function(pattern = NULL,
+                            n_max = Inf,
+                            ...,
+                            verbose = TRUE) {
+  lifecycle::deprecate_warn("2.0.0", "team_drive_find()", "shared_drive_find()")
+  # TODO: add something about verbosity once I deprecate that
+  shared_drive_find(
+    pattern = pattern,
+    n_max = n_max,
+    ...,
+    verbose = verbose
+  )
+}
+
+#' @export
+#' @rdname deprecated-team-drive-functions
+team_drive_get <- function(name = NULL, id = NULL, verbose = TRUE) {
+  lifecycle::deprecate_warn("2.0.0", "team_drive_get()", "shared_drive_get()")
+  # TODO: add something about verbosity once I deprecate that
+  shared_drive_get(name = name, id = id, verbose = verbose)
+}
+
+#' @export
+#' @rdname deprecated-team-drive-functions
+team_drive_create <- function(name, verbose = TRUE) {
+  lifecycle::deprecate_warn("2.0.0", "team_drive_create()", "shared_drive_create()")
+  # TODO: add something about verbosity once I deprecate that
+  shared_drive_create(name = name, verbose = verbose)
+}
+
+#' @export
+#' @rdname deprecated-team-drive-functions
+team_drive_rm <- function(team_drive = NULL, verbose = TRUE) {
+  lifecycle::deprecate_warn("2.0.0", "team_drive_rm()", "shared_drive_rm()")
+  # TODO: add something about verbosity once I deprecate that
+  shared_drive_rm(drive = team_drive, verbose = verbose)
+}
+
+#' @export
+#' @rdname deprecated-team-drive-functions
+team_drive_update <- function(team_drive, ..., verbose = TRUE) {
+  lifecycle::deprecate_warn("2.0.0", "team_drive_update()", "shared_drive_update()")
+  # TODO: add something about verbosity once I deprecate that
+  shared_drive_update(drive = team_drive, ..., verbose = verbose)
+}
+
+#' @export
+#' @rdname deprecated-team-drive-functions
+ as_team_drive <- function(x, ...) {
+  lifecycle::deprecate_warn("2.0.0", "as_team_drive()", "as_shared_drive()")
+  as_shared_drive(x, ...)
+ }
+
+#' @export
+#' @rdname deprecated-team-drive-functions
+is_team_drive <- function(d) {
+  stopifnot(inherits(d, "dribble"))
+  purrr::map_chr(d$drive_resource, "kind") == "drive#teamDrive"
+}