+++ man-roxygen/shared-drive-description.R
@@ -1,0 +1,7 @@
+#' @description
+
+#' A shared drive supports files owned by an organization rather than an
+#' individual user. Shared drives follow different sharing and ownership models
+#' from a specific user's "My Drive". Shared drives are the successors to the
+#' earlier concept of Team Drives. Learn more about [shared
+#' drives][shared_drives].