+++ man-roxygen/team-drives-description.R
@@ -1,4 +1,13 @@
-#' @description Note: [Team
-#'   Drives](https://gsuite.google.com/learning-center/products/drive/get-started-team-drive/)
-#'    are only available to users of certain enhanced Google services, such as G
-#'   Suite Enterprise, G Suite Business, or G Suite for Education.
+#' @description
+
+#' Team Drives have been rebranded as *shared drives* and, as of googledrive
+#' v2.0.0, all `team_drive_*()` functions have been deprecated, in favor of
+#' `shared_drive_*()` successors.
+#'
+#' The changes in googledrive reflect that the Team Drives resource collection
+#' has been deprecated in the Drive API v3, in favor of the new (shared) Drives
+#' resource collection. Read more
+
+#' * <https://cloud.google.com/blog/products/application-development/upcoming-changes-to-the-google-drive-api-and-google-picker-api>
+
+